refactor: 패키지 분리 및 기본 접근 지정자 사용
