+++ src/main/java/blackjack/domain/game/Result.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.game;
 
+import blackjack.domain.participant.Player;
 import java.util.Map;
 
 public class Result {
