+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -1,5 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
+import blackjack.domain.game.Result;
+import blackjack.domain.game.ResultStatus;
 import java.util.LinkedHashMap;
 import java.util.Map;
