+++ src/test/java/blackjack/domain/ResultTest.java
@@ -3,6 +3,9 @@ package blackjack.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import blackjack.domain.game.Result;
+import blackjack.domain.participant.Betting;
+import blackjack.domain.participant.Player;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
