+++ src/test/java/blackjack/domain/participant/DealerTest.java
@@ -1,8 +1,14 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardSuit;
+import blackjack.domain.card.Deck;
+import blackjack.domain.fixture.CardFixture;
+import blackjack.domain.game.Result;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
